feat(check): dead metrics
Check for outdated metrics and more sanity.
This check is fine to run and does not really bother the databases because it only uses the most recent value.
The previous check for infinite values is much tougher because it forces all databases to open all files, and nfs does not like that.

Fixes #14